refactor: move block-thumbnails into adapter-gui (#135)
Remove the standalone block-thumbnails crate and inline its logic
into adapter-gui as a thumbnails module. The build.rs thumbnail
generation now runs as part of adapter-gui's build step.

- Move lib.rs → adapter-gui/src/thumbnails.rs
- Merge build.rs thumbnail generation into adapter-gui/build.rs
- Remove block-thumbnails from workspace members
- Remove block-thumbnails dependency from adapter-gui
