+++ crates/adapter-gui/Cargo.toml
@@ -7,7 +7,6 @@ edition.workspace = true
 anyhow.workspace = true
 application = { path = "../application" }
 block-core = { path = "../block-core" }
-block-thumbnails = { path = "../block-thumbnails" }
 image = { version = "0.25", default-features = false, features = ["png"] }
 cpal = "0.17"
 domain = { path = "../domain" }
