+++ crates/adapter-gui/src/thumbnails.rs
