--- crates/block-thumbnails/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "block-thumbnails"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
